requirements.txt
arrow==1.3.0
asgiref==3.8.1
binaryornot==0.4.4
cachetools==5.5.0
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cookiecutter==2.6.0
Django==4.2.16
django-browser-reload==1.15.0
django-tailwind==3.6.0
google-api-core==2.23.0
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.66.0
httplib2==0.22.0
idna==3.8
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
oauthlib==3.2.2
proto-plus==1.25.0
protobuf==5.29.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-slugify==8.0.4
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.8.1
rsa==4.9
six==1.16.0
sqlparse==0.5.1
text-unidecode==1.3
types-python-dateutil==2.9.0.20240906
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
